requirements.txt
astroid==2.4.2
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
cryptography==3.0
decorator==4.4.2
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
lxml==4.5.2
mccabe==0.6.1
pycparser==2.20
pylint==2.5.3
PySocks==1.7.1
python-telegram-bot==12.8
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
toml==0.10.1
tornado==6.0.4
urllib3==1.25.10
wrapt==1.12.1
